Support for loading from Storyboards and NIBs
The previous code would not work when assigning a UIView in InterfaceBuilder or the Storyboard template to the tab view. The method initWithFrame will not get called with using IB or Storyboards and therefore the tab view will not appear. Moved the setup code in its own function that will be called regardless of method being used. In addition, made the `delegate` property accessible through IB.